requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysql-connector-python
django-filter
drf-spectacular
python-decouple
pytest
pytest-django
boto3
django-storages
gunicorn
psycopg2-binary
dj-database-url
whitenoise

# Code Quality & Formatting
black
isort
flake8

# Security Scanning
bandit
safety

# Dead Code Detection
vulture

# Testing Tools
factory-boy
pytest-cov

# Pre-commit Hooks
pre-commit
